refactor program and run func
